configurations with main menu and main activity
